Compose and multiple networks (#20)
* nit! go fmt changes

...broken out to make it easier to see later changes.

* Resolve names against compose labels

docker-compose applies a number of well-known labels to help identify
containers.  These labels would seem to be useful for service discovery.

This change creates a new resolver, ComposeResolver, and adds a
configuration option (compose_hostname) to allow it to be enabled.

* Allow specific networks to be specified

It's not horribly unusual for containers to have multiple addresses --
that is, be on multiple networks.  This change allows us to specify a
specific network to use when selecting an address.

* Don't use top-level IP when specific network given

I'm not 100% sure when IPAddress in NetworkSettings is populated, but if
we've specified a network via a label it stands to reason that we'll
want to ignore it in favor of the ip on the specific network.

* Add a note on `compose_domain` to the README

* Update README.md

Co-authored-by: Francois Ferrand <thetypz@gmail.com>

Co-authored-by: Francois Ferrand <thetypz@gmail.com>